pyproject.toml
[project]
name = "story-writer"
version = "0.1.0"
description = "Oda-Style Manga Story Engine - Generate infinite serialized manga narratives"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # LLM clients
    "anthropic>=0.39.0",
    "openai>=1.54.0",

    # Data validation and models
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",

    # Configuration and utilities
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",

    # Storage (Phase 2 - optional for now)
    # "sqlalchemy>=2.0.0",
    # "chromadb>=0.4.0",
    # "sentence-transformers>=2.2.0",

    # CLI (Phase 4 - optional for now)
    # "typer>=0.9.0",
    # "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
story-writer = "story_writer.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_backend"
